% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{average_hierarchy}
\alias{average_hierarchy}
\title{Plot the average dendrogram for the replicate optimization}
\usage{
average_hierarchy(replicates)
}
\arguments{
\item{replicates}{A list of results optimization replicates from best_clusters.}
}
\value{
The dendrogram obtained from the simulation results for each lambda
}
\description{
Plot the average dendrogram for the replicate optimization
}
