% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operation.R
\name{build_theta}
\alias{build_theta}
\title{From R matrix compute theta matrix}
\usage{
build_theta(R, clusters)
}
\arguments{
\item{R}{a K x K matrix : the matrix of the clusters coefficients}

\item{clusters}{a list of vector : each vector gives the element of a
cluster.}
}
\value{
Return the theta matrix from the value of the R matrix of clusters
coefficient using :

\if{html}{\out{<div class="sourceCode">}}\preformatted{                       theta = U R U^t + A
}\if{html}{\out{</div>}}

where U is the cluster matrix and a diagonal matrix such that the rows of
theta sum to zero :
a_ii = - sum_l p_l r_kl
for i in cluster C_k.
}
\description{
From R matrix compute theta matrix
}
\examples{

R <- matrix(c(1, 2,
              2, 5), nr = 2)
clusters <- list(c(1,2,3),c(4,5))
build_theta(R, clusters)

}
