% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{compare_clusters}
\alias{compare_clusters}
\title{Compare two clusters.}
\usage{
compare_clusters(clusters1, clusters2)
}
\arguments{
\item{clusters1}{The first cluster to compare.}

\item{clusters2}{The second cluster to compare.}
}
\value{
Return FALSE if the clusters have not the same number of cluster. Otherwise,
}
\description{
Compare two clusters.
}
\keyword{internal}
