% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operation.R
\name{compute_W}
\alias{compute_W}
\title{Computation of the first weight matrix}
\usage{
compute_W(data)
}
\arguments{
\item{data}{the data}
}
\value{
The initial wieght matrix without fused variables.
}
\description{
Computation of the first weight matrix
}
\keyword{internal}
