% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operation.R
\name{crout_factorisation}
\alias{crout_factorisation}
\title{Crout factorization algorithm.}
\usage{
crout_factorisation(A, tol = 1e-12)
}
\arguments{
\item{A}{a d x d symmetric positive matrix.}

\item{tol}{a positive value : tolerance for the zero diagonal}
}
\value{
The LU Crout decomposition of a matrix A. For A a symmetric
positive matrix, there exists a LU decomposition such that :
A = L' L
}
\description{
Crout factorization algorithm.
}
\examples{
A <- matrix(c(1,2,3,
             2,5,6,
             3,6,9), nc = 3)
L <- crout_factorisation(A)
L \%*\% t(L)
}
\keyword{internal}
