% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operation.R
\name{D_tilde2_r}
\alias{D_tilde2_r}
\title{Cluster distance squared function}
\usage{
D_tilde2_r(R, clusters)
}
\arguments{
\item{R}{K x K symmetric matrix.}

\item{clusters}{a list of vector : each vector gives the element of
a cluster.}
}
\value{
A function of cluster number : compute the square distance between
two clusters for the distance defined in section 4.2 in cluster document.
}
\description{
Cluster distance squared function
}
\examples{

R <- matrix(c(0.5, -1,
              -1, -1), nr = 2)
clusters <- list(c(1,3), c(2,4))
D2 <- D_tilde2_r(R, clusters)
D2(1, 2)
}
\keyword{internal}
