% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{detect_merge}
\alias{detect_merge}
\title{Detect the merge and give all information about cluster and lambda.}
\usage{
detect_merge(solution_list)
}
\arguments{
\item{solution_list}{Solution's list from best_clusters.}
}
\value{
Return the list of the solution where a change occurs between the list of clusters.
}
\description{
Detect the merge and give all information about cluster and lambda.
}
\keyword{internal}
