% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrc_refit.R
\name{.HRC_refit_wrapper}
\alias{.HRC_refit_wrapper}
\title{Refit a Hüsler-Reiss model with a given clustering and variogram matrix}
\usage{
.HRC_refit_wrapper(clusters, Gamma)
}
\description{
Refit a Hüsler-Reiss model with a given clustering and variogram matrix
}
\keyword{internal}
