% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{.HRC_wrapper}
\alias{.HRC_wrapper}
\title{Wrapper function for the HRClusterpath procedure implemented in .HRClusterpath (in C++).}
\usage{
.HRC_wrapper(
  data,
  zeta,
  lambda,
  p,
  W,
  kappa,
  eps_conv,
  eps_f,
  tol_opt,
  iter_max
)
}
\description{
Wrapper function for the HRClusterpath procedure implemented in .HRClusterpath (in C++).
}
\keyword{internal}
