% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operation.R
\name{extract_R_matrix}
\alias{extract_R_matrix}
\title{Function to extract the coefficient of the reduced matrix R}
\usage{
extract_R_matrix(matrix, clusters)
}
\arguments{
\item{matrix}{A block matrix which can be factorizable.}

\item{clusters}{a list of vectors : the variable index per clusters.}
}
\value{
A matrix on size the number of clusters (length(clusters)). It
corresponds to the reduced matrix R of the original matrix (see clusters document
for definition).)
}
\description{
Function to extract the coefficient of the reduced matrix R
}
