% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operation.R
\name{gamma_function}
\alias{gamma_function}
\title{Variogram transformation application gamma}
\usage{
gamma_function(sigma)
}
\arguments{
\item{sigma}{A d x d numeric matrix.}
}
\value{
For a symmetric positive matrix sigma (covariance matrix),
return the corresponding variogram matrix. Can be used for other
but with no interpretation.
}
\description{
Variogram transformation application gamma
}
\examples{
s_sigma <- matrix(rnorm(16, 2), nc = 4)
gamma_function(s_sigma \%*\% t(s_sigma))
}
\keyword{internal}
