% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operation.R
\name{gen_det}
\alias{gen_det}
\title{Generalised determinant}
\usage{
gen_det(A, tol = 1e-10)
}
\arguments{
\item{A}{a d x d real valued matrix.}

\item{tol}{a positive value.}
}
\value{
Compute the generalised determinant of a matrix A. We recall
that the generalised determinant is an extension of the determinant for
singular matrix. It corresponds to the product of all the non zero eigen
values.
}
\description{
Generalised determinant
}
\examples{
A <- matrix(c(1,2,3,
              2,5,6,
              3,6,9), nc = 3)
gen_det(A)
}
\keyword{internal}
