% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{get_adjacency_matrix}
\alias{get_adjacency_matrix}
\title{Give the adjacency matrix coresponding to the right hierarchical according to the fusion
during optimization.}
\usage{
get_adjacency_matrix(event_list, lambda_max)
}
\arguments{
\item{event_list}{the list of solution from detect_fusion.}

\item{lambda_max}{The lambda which end the optimization.}
}
\value{
A matrix which will be use for the construction of the dendrogram.
}
\description{
Give the adjacency matrix coresponding to the right hierarchical according to the fusion
during optimization.
}
\keyword{internal}
