% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{get_cluster}
\alias{get_cluster}
\title{Gradient descent algorithm for Husler-Reiss graphical models clustering}
\usage{
get_cluster(gamma, weights, eps_f, ...)
}
\arguments{
\item{gamma}{a d x d matrix : the variogram matrix.}

\item{weights}{a d x d symmetric matrix with a zero diagonal.}

\item{...}{}

\item{lambda}{a positive number : the weight of the penalty.}
}
\value{
A function which compute the maximum likelihood estimator using
cluster-path gradient descent and returns the estimation of the clusters and
the corresponding R matrix.
}
\description{
Gradient descent algorithm for Husler-Reiss graphical models clustering
}
\examples{
W <- matrix(c(0, 1, 1, 1,
              1, 0, 1, 1,
              1, 1, 0, 1,
              1, 1, 1, 0), nc = 4)
gamma <- generate_random_Gamma(d = 4)
R <- matrix(c(1,0,0,-1,
              0,1,1,-2,
              0,1,1,-1,
              -1,-2,-1,1), nc = 4)
Cluster_HR <- get_cluster(gamma, W, 100)
Cluster_HR(R)

}
