% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{gg_cluster}
\alias{gg_cluster}
\title{Plot the dendrogram for the optimization}
\usage{
gg_cluster(list_results)
}
\arguments{
\item{list_results}{A list of results optimization from best_clusters.}
}
\value{
The dendrogram obtained from the simulation results for each lambda.
}
\description{
Plot the dendrogram for the optimization
}
