% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{ggdistance}
\alias{ggdistance}
\title{Graphicall representation of Clusterpath results with multidimensional
scaling.}
\usage{
ggdistance(list_results, names)
}
\arguments{
\item{list_results}{A list of results optimization from \code{\link{HR_Clusterpath}()}.}

\item{names}{}
}
\description{
Graphicall representation of Clusterpath results with multidimensional
scaling.
}
