% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{gradient_D2}
\alias{gradient_D2}
\title{Gradient matrix of distance between two columns}
\usage{
gradient_D2(R, clusters)
}
\arguments{
\item{R}{K x K symmetric matrix.}

\item{clusters}{a list of vector : each vector gives the element of
a cluster.}
}
\value{
Return a functon of indices (k', l') computing the gradient
matrix of tilde D^2(r_k', r_l'). See section 4.3.3 in cluster document
for details.
}
\description{
Gradient matrix of distance between two columns
}
\examples{
R <- matrix(c(-1,0,-2,
              0,-3,-1,
              -2,-1,-1), 3)
clusters <- list(c(1,3), c(2), 4)
grad <- gradient_D2(R, clusters)
grad(1, 2)
}
\keyword{internal}
