% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrc_refit.R
\name{hr-clusterpath-refit}
\alias{hr-clusterpath-refit}
\alias{HR_Clusterpath_refit}
\title{H端sler-Reiss Clusterpath refitted}
\usage{
HR_Clusterpath_refit(hrc_output)
}
\arguments{
\item{hrc_output}{An object of the \code{HR_Clusterpath} function.}
}
\value{
An object of class \code{HR_Clusterpath_refit} with the same structure as the \code{HR_Clusterpath}
output but with refitted R matrices for each result.
}
\description{
The refit procedure is based on a convex optimization problem which is solved with the \code{CVXR} package.
Based on the results produced by \code{HR_Clusterpath}, a new problem is defined as the minimization of the
negative loglikelihood of a H端sler-Reiss model with precision matrix \eqn{\Theta} under the constraints
of block matrix structure induced by the estimated clusters. These solutions are computed for each value of
\eqn{\lambda} in the output of \code{HR_Clusterpath}.
}
\section{Motivation }{

The problem solution provided by \code{HR_clusterpath} may be biased due to the penalty in the
negative loglikelihood. To avoid this potential issue, the matrix \eqn{R} is refitted for
each solution over \eqn{\lambda} with the non-penalized negative loglikelihood constraining by the
estimated clusters.
}

\examples{
# Initialization of the parameters for the simulation
R <- matrix(c(1, -3, 0,
              -3, 2, -2,
              0, -2, 1), nc = 3)
clusters <- list(1:5, 6:10, 11:15)

# Building of the variogram matrix for the simulation
Theta <- build_Theta(R, clusters)
Gamma <- graphicalExtremes::Theta2Gamma(Theta)

# Simulation of data
set.seed(123)
n <- 1e4
d <- ncol(Gamma)
data <- graphicalExtremes::rmstable(
           n = n,
           model = "HR",
           d = d,
           par = Gamma)

# Computation of the multivariate Pareto data
norm_inf <- apply(abs(data), 1, max)
quantile <- sort(norm_inf, decreasing = TRUE) |> as.numeric()
k <- floor(0.1 * nrow(data))
u <- quantile[k]

data_par <- data[norm_inf > u, ] / u         # Multivariate Pareto data

# Computation of the H端sler-Reiss Clusterpath
zeta <- log(n) ** 2              # Zeta parameter for the exponential weights
HRC <- HR_Clusterpath(data_par, lambda = 10, zeta = zeta)

# First estimation of R
HRC$results$l1$R

# Refit of the H端sler-Reiss Clusterpath
HRC_refit <- HR_Clusterpath_refit(HRC)

# New estimation of R
HRC_refit$results$l1$R
}
