% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrc_refit.R
\name{HR_Clusterpath_refit}
\alias{HR_Clusterpath_refit}
\title{Refit the HR_clusterpath output for each results}
\usage{
HR_Clusterpath_refit(hrc_output)
}
\arguments{
\item{hrc_output}{An output of the \code{HR_Clusterpath} function.}
}
\value{
An output of class \code{HR_Clusterpath_refit} with the same structure as the \code{HR_Clusterpath}
output but with refitted R matrices for each result.
}
\description{
Refit the HR_clusterpath output for each results
}
