% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrc_hierarchy.R
\name{hrc-hierarchy}
\alias{hrc-hierarchy}
\alias{HR_Clusterpath_hierarchy}
\title{Clusterpath algorithm for Hüsler-Reiss models hierarchical version}
\usage{
HR_Clusterpath_hierarchy(
  Gamma,
  lambda,
  W = NULL,
  zeta,
  eps_f = NULL,
  kappa = 0.01,
  eps_conv = 1e-07,
  tol_opt = 0.001,
  iter_max = 1000
)
}
\arguments{
\item{Gamma}{A \eqn{d \times d} matrix : a variogram matrix.}

\item{lambda}{A numeric vector of positive number : the grid line for \eqn{\lambda}.}

\item{W}{A \eqn{d \times d} matrix of positive number : the weights for the penalty.
If \code{NULL} (default), it is set as the exponential weights, which depends of only one
tuning parameter \eqn{\zeta} and where we have some theoretical results.}

\item{zeta}{A positive number : tuned parameter for the exponential weights.
If \code{W} is not \code{NULL}, this parameter is not used.}

\item{eps_f}{A positive number : tolerance threshold for merging clusters. If \code{NULL} (default),
it is set as \eqn{\kappa} times the median of the non-diagonal elements of the distance matrix
between the column of the initial \eqn{\Theta}.}

\item{kappa}{A positive number : tuned parameter for the fusion threshold. If \code{eps_f} is not \code{NULL},
this parameter is not used.}

\item{eps_conv}{A positive number : tolerance threshold for the convergence of the algorithm.}

\item{tol_opt}{A positive number : tolerance for the optimal step in the gradient descent.}

\item{iter_max}{An integer : maximal number of iterations of the procedure.}
}
\value{
\code{HR_Clusterpath()} is a block gradient descent from the data with default values and method for
the optimization : the variogram matrix \eqn{\Gamma} is the empirical variogram and the weights are
set as the exponential weights, which depends of only one tuning parameter \eqn{\zeta} and where
some theoretical guarantees are provided.

The output is an object of class \code{HRC_hierarchy} which contains a list of several datas :
\itemize{
\item $results : a list of results for each value of lambda, where each result is a
list with the following elements : $R the \eqn{R} matrix of the clusters, $clusters
a list of the variable indices separated per cluster, $likelihood the value of the
penalised negative loglikelihood and $lambda the corresponding regularized parameter.
\item $Gamma : the variogram matrix used for the procedure.
\item $inputs : a list of the input parameters used in the procedure.
}
}
\description{
Block gradient descent based on the Clusterpath algorithm for Hüsler-Reiss
graphical model. Usefull when the precision matrix \eqn{\Theta}
(or the variogram \eqn{\Gamma}) has a block matrix structure.
}
\details{
The block matrix models are characterized by two objects : a clustering partition
\eqn{\{C_1, \dots, C_K\}} of \eqn{V = \{1, \dots, d\}} and a matrix \eqn{R}
which belongs to \eqn{\mathcal S_K(\mathbb R)}, the set of symmetric \eqn{K \times K}
matrix (See also \code{\link{build_theta}()} and \code{\link{extract_R_matrix}()}).

The Clusterpath aims to find the optimum of some penalised negative loglikelihood defined in
\code{\link{neg_likelihood_pen}()}. The path of solution
is computed using the Clusterpath procedure on a mesh for \eqn{\lambda} where each
initialization is provided using the results of the previous \eqn{\lambda} within the grid.
}
\examples{
# Initialization of the parameters for the simulation
R <- matrix(c(1, -3, 0,
              -3, 2, -2,
              0, -2, 1), nc = 3)
clusters <- list(1:5, 6:10, 11:15)

# Building of the variogram matrix for the simulation
Theta <- build_Theta(R, clusters)
Gamma <- graphicalExtremes::Theta2Gamma(Theta)

# Simulation of data
set.seed(123)
n <- 1e4                               # Sample size
d <- ncol(Gamma)                       # Number of variables
data <- graphicalExtremes::rmstable(
           n = n,
           model = "HR",
           d = d,
           par = Gamma)

# Computation of the multivariate Pareto data
norm_inf <- apply(abs(data), 1, max)
quantile <- sort(norm_inf, decreasing = TRUE) |> as.numeric()
k <- floor(0.1 * nrow(data))
u <- quantile[k]

data_par <- data[norm_inf > u, ] / u        # Multivariate Pareto data

# Computation of the Hüsler-Reiss Clusterpath
zeta <- log(n) ** 2              # Zeta parameter for the exponential weights
Gamma_est <- graphicalExtremes::emp_vario(data_par) # Estimated variogram
lambda <- seq(0, 1, 0.1)                            # Meshgrid for lambda

HRC <- HR_Clusterpath_hierarchy(Gamma_est, lambda = lambda, zeta = zeta)
HRC$results$l10
}
