% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{merge_clusters}
\alias{merge_clusters}
\title{Function which merges clusters}
\usage{
merge_clusters(R, clusters, eps = 0.1, cost)
}
\arguments{
\item{R}{K x K symmetric matrix.}

\item{clusters}{a list of vector : each vector gives the element of
a cluster.}

\item{eps}{positive value : minimal tolerance for merging clusters}

\item{cost}{a function : Cost function of the optimisation}
}
\value{
Returns, if merging, a list of the new clusters and the
corresponding R matrix, where the coefficient of the new clustered
is computing by averaging the coefficient of the two previous clusters.
}
\description{
Function which merges clusters
}
\examples{
R <- matrix(c(0.5, -1,
              -1, -1), nr = 2)
clusters <- list(c(1,3), c(2,4))
W <- matrix(c(0, 1, 1, 1,
              1, 0, 1, 1,
              1, 1, 0, 1,
              1, 1, 1, 0), nc = 4)
gamma <- matrix(c(0,2,1,0,
                  2,0,4,1,
                  1,4,0,7,
                  0,1,7,0), nc = 4)
cost <- neg_likelihood_pen(gamma, weights, 100000)
merge_clusters(R, clusters, cost = cost)

}
\keyword{internal}
