% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{multi_scale_data}
\alias{multi_scale_data}
\title{Build the dataframe for multidimensional scaling for Clusterpath.}
\usage{
multi_scale_data(list_res)
}
\arguments{
\item{list_res}{the list of solution from \code{HR_Clusterpath}.}
}
\value{
A matrix containing the projection of the variable results on the
first direction and the associated value of lambda.
}
\description{
Build the dataframe for multidimensional scaling for Clusterpath.
}
\keyword{internal}
