% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{neg_likelihood}
\alias{neg_likelihood}
\title{Negative log-likelihood computation.}
\usage{
neg_likelihood(gamma)
}
\arguments{
\item{gamma}{A \eqn{d \times d} matrix: the empirical variogram matrix.}
}
\value{
For a fixed variogram gamma, compute for a set of clusters and
corresponding \eqn{R} matrix, the value of the associated negative likelihood
defined by :

\deqn{
     L(\Theta) = - \log(|\Theta|_+) - \frac 1 2 \text{tr}(\hat \Gamma \Theta)
}

where \eqn{|\cdot|_+} is the generalised determinant.
}
\description{
Negative log-likelihood computation.
}
\examples{
R <- matrix(c(0.5, -1,
              -1, -1), nr = 2)
clusters <- list(c(1,3), c(2,4))
gamma <- matrix(c(0, 2, 1, 0,
                  2, 0, 4, 1,
                  1, 4, 0, 7,
                  0, 1, 7, 0), nc = 4)
nllh <- neg_likelihood(gamma)
nllh(R, clusters)

}
