% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{neg_likelihood_pen}
\alias{neg_likelihood_pen}
\title{Computation of the penalised negative log-likelihood}
\usage{
neg_likelihood_pen(gamma, weights)
}
\arguments{
\item{gamma}{a d x d matrix : the variogram matrix.}

\item{weights}{a d x d symmetric matrix with a zero diagonal.}

\item{lambda}{a positive number : the weight of the penalty.}
}
\value{
A function of clusters and the R matrix which compute the penalised
negative log-likelihood of the model.
}
\description{
Computation of the penalised negative log-likelihood
}
\examples{
R <- matrix(c(0.5, -1,
              -1, -1), nr = 2)
clusters <- list(c(1,3), c(2,4))
W <- matrix(c(0, 1, 1, 1,
              1, 0, 1, 1,
              1, 1, 0, 1,
              1, 1, 1, 0), nc = 4)
gamma <- matrix(c(0,2,1,0,
                  2,0,4,1,
                  1,4,0,7,
                  0,1,7,0), nc = 4)
f <- neg_likelihood_pen(gamma, W, 0.5)
f(R, clusters)

}
