% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{nloglike_grad_np}
\alias{nloglike_grad_np}
\title{Gradient of the negative log likelihood without penalty}
\usage{
nloglike_grad_np(gamma)
}
\arguments{
\item{gamma}{a d x d variogram matrix.}
}
\value{
A function of the R matrix and clusters and compute the gradient
matrix of the negative log likelihood for a fixed variogram gamma. The
gradient matrix can be computed by :

\if{html}{\out{<div class="sourceCode">}}\preformatted{                   dnllh = dlog + dtrace
}\if{html}{\out{</div>}}

where :
- dlog = t(U) g(Theta_p) U - 0.5 diag(t(U) g(Theta_p) U)
- dtrace = - t(U) gamma U  + 0.5 diag(t(U) gamma U)
}
\description{
Gradient of the negative log likelihood without penalty
}
\examples{
R <- matrix(c(0.5, -1,
              -1, -1), nr = 2)
clusters <- list(c(1,3), c(2,4))
gamma <- matrix(c(0,2,1,0,
                  2,0,4,1,
                  1,4,0,7,
                  0,1,7,0), nc = 4)
gradient <- nloglike_grad_np(gamma)
gradient(R, clusters)

}
\keyword{internal}
