% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{nloglike_grad_np_gamma}
\alias{nloglike_grad_np_gamma}
\title{Gradient of the negative log likelihood without penalty for inverse problem.}
\usage{
nloglike_grad_np_gamma(Theta)
}
\arguments{
\item{Theta}{a d x d precision matrix}
}
\value{
A function of the  clusters and compute the gradient
matrix of the negative log likelihood for a fixed Precision matrix Theta.
}
\description{
Gradient of the negative log likelihood without penalty for inverse problem.
}
\examples{
clusters <- list(c(1,3), c(2,4))
Theta <- matrix(c(3,-2,-1,0,
                  -2,7,-4,-1,
                  -1,-4,12,-7,
                  0,-1,-7,8), nc = 4)
gradient <- nloglike_grad_np(Theta)
gradient(clusters)

}
