% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{penalty}
\alias{penalty}
\title{Penalty function.}
\usage{
penalty(weights)
}
\arguments{
\item{weights}{a \eqn{d \times d} symmetric matrix with a zero diagonal.}
}
\value{
For fixed weights, returns a function which compute the value of
the penalty for chosen clusters and corresponding R matrix. We recall the
penalty is given by :
P(R) = sum_{k<l} W_kl D^2(r_.k, r_.l)
with W_kl the clustered weight for clusters k and l.
}
\description{
Penalty function.
}
\examples{
R <- matrix(c(0.5, -1,
              -1, -1), nr = 2)
clusters <- list(c(1,3), c(2,4))
W <- matrix(c(0, 1, 1, 1,
              1, 0, 1, 1,
              1, 1, 0, 1,
              1, 1, 1, 0), nc = 4)
P <- penalty(W)
P(R, clusters)
}
