% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{penalty_grad}
\alias{penalty_grad}
\title{Computation of the penalty's gradient}
\usage{
penalty_grad(weights)
}
\arguments{
\item{weights}{a d x d symmetric matrix with a zero diagonal.}
}
\value{
A function of clusters and corresponding R matrix. Compute
the gradient with fixed weight. The expression of the gradient is
just the weighted sum of the gradient of each tilde D^2 where the
weights are the clustered weights.
See equations in section 4.3.3 for details.
}
\description{
Computation of the penalty's gradient
}
\examples{
R <- matrix(c(0.5, -1,
              -1, -1), nr = 2)
clusters <- list(c(1,3), c(2,4))
W <- matrix(c(0, 1, 1, 1,
              1, 0, 1, 1,
              1, 1, 0, 1,
              1, 1, 1, 0), nc = 4)
dpen <- penalty_grad(W)
dpen(R, clusters)
}
\keyword{internal}
