% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operation.R
\name{semi_def}
\alias{semi_def}
\title{Semi-definite checker}
\usage{
semi_def(M)
}
\arguments{
\item{M}{A symmetric matrix.}
}
\value{
TRUE if the matrix is semi-definite positive, FALSE otherwise. The
verification is done by checking the sign of the eigen-values of the matrix.
}
\description{
Semi-definite checker
}
\keyword{internal}
