% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{step_gradient}
\alias{step_gradient}
\title{Step for the gradient descent}
\usage{
step_gradient(gamma, weights, size_grid = 100)
}
\arguments{
\item{gamma}{a d x d matrix : the variogram matrix.}

\item{weights}{a d x d symmetric matrix with a zero diagonal.}

\item{size_grid}{integer : size of the search grid for the optimal step.}

\item{lambda}{a positive number : the weight of the penalty.}
}
\value{
A function of clusters and R matrix which returns the next step of
the optimisation for the gradient descent algorithm.
}
\description{
Step for the gradient descent
}
\examples{
R <- matrix(c(0.5, -1,
              -1, -1), nr = 2)
clusters <- list(c(1, 3), c(2,4))
W <- matrix(c(0, 1, 1, 1,
              1, 0, 1, 1,
              1, 1, 0, 1,
              1, 1, 1, 0), nc = 4)
gamma <- matrix(c(0, 2, 1, 0,
                  2, 0, 4, 1,
                  1, 4, 0, 7,
                  0, 1, 7, 0), nc = 4)
f <- step_gradient(gamma, W, 0.5)
f(R, clusters)

}
\keyword{internal}
