% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operation.R
\name{theta-r}
\alias{theta-r}
\alias{build_theta}
\alias{extract_R_matrix}
\title{Functions to navigate between \eqn{R} and \eqn{\Theta} for given clusters}
\usage{
build_theta(R, clusters)

extract_R_matrix(Theta, clusters)
}
\arguments{
\item{R}{For \code{build_theta()}, the \eqn{K \times K} matrix of the clusters.}

\item{clusters}{A list of indices associated to a partition of \eqn{V}.}

\item{Theta}{For \code{extract_R_matrix()}, a \eqn{d \times d} block matrix which can be factorizable.}
}
\value{
For \code{extract_R_matrix()}, a matrix on size the number of clusters (i.e. \eqn{K})
corresponding to the reduced matrix \eqn{R} of the original matrix \eqn{\Theta}.

For \code{build_theta()}, a matrix on size the number of variables (i.e. \eqn{d})
corresponding to the precision matrix \eqn{\Theta} induced by \eqn{R}.
}
\description{
The block matrix models is defined from two elements : the cluster's partition
\eqn{\{C_1, \dots, C_K\}}, included in \eqn{V} and the \eqn{R} matrix which belongs
to \eqn{\mathcal S_K(\mathbb R)}, the set of symmetric \eqn{K \times K} matrix.
The expression of the precision \eqn{\Theta} which statisfies the block matrix model is
\deqn{
   \Theta = U R U^t + A
}
where \eqn{U} is the cluster matrix and \eqn{A} a diagonal matrix such that the rows of
theta sum to zero :
\eqn{
a_{ii} = - \sum_l p_l r_{kl}
}
for \eqn{i} in cluster \eqn{C_k}.
}
\examples{
##############################################################
#                       FROM THETA TO R
##############################################################
Theta <- matrix(
 c(4.5, .5, .5, .5, -2, -2, -2,
   .5, 4.5, .5, .5, -2, -2, -2,
   .5, .5, 4.5, .5, -2, -2, -2,
   .5, .5, .5, 4.5, -2, -2, -2,
   -2, -2, -2, -2, 6, 1, 1,
   -2, -2, -2, -2, 1, 6, 1,
   -2, -2, -2, -2, 1, 1, 6),
 nc = 7
)

clusters <- list(c(1,2,3,4), c(5,6,7))

extract_R_matrix(Theta, clusters)

##############################################################
#                       FROM R TO THETA
##############################################################

R <- matrix(c(1, -3, 0,
              -3, 2, -2,
              0, -2, 1), nc = 3)

clusters <- list(1:4, 5:8, 9:12)

build_theta(R, clusters)


}
