% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operation.R
\name{U_matrix}
\alias{U_matrix}
\title{Computation of the matrix of clusters}
\usage{
U_matrix(clusters)
}
\arguments{
\item{clusters}{a list of vector : each vector gives the element of a
cluster.}
}
\value{
The d x K matrix U such that u_jk = 1 if the variable j belongs to
cluster C_k and 0 otherwise.
}
\description{
Computation of the matrix of clusters
}
\examples{
clusters <- list(c(1,2,3),c(4,5))
U_matrix(clusters)

}
\keyword{internal}
