% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operation.R
\name{weight_clustered}
\alias{weight_clustered}
\title{Computation of the clustered weight matrix}
\usage{
weight_clustered(weights)
}
\arguments{
\item{weights}{a d x d symmetric matrix with a zero diagonal.}
}
\value{
A function of clusters : the weight matrix for each pair
of clusters :
W_kl = sum_(i in C_k) sum_(j in C_l) w_ij
}
\description{
Computation of the clustered weight matrix
}
\examples{
clusters <- list(c(1,2,3), c(4,5))
W <- matrix(1:25, nc = 5)
W_c <- weight_clustered(W)
W_c(clusters)
}
\keyword{internal}
