% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operation.R
\name{ChiToGamma}
\alias{ChiToGamma}
\title{Transform a chi matrix to the corresponding variogram}
\usage{
ChiToGamma(Chi_matrix)
}
\arguments{
\item{Chi_matrix}{the matrix with the chi coefficient}
}
\value{
The corresponding variogram matrix $\Gamma$.
}
\description{
Transform a chi matrix to the corresponding variogram
}
