% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operation.R
\name{extract_R_matrix}
\alias{extract_R_matrix}
\title{Function to extract the coefficients of the reduced matrix R.}
\usage{
extract_R_matrix(Theta, clusters)
}
\arguments{
\item{Theta}{A block matrix which can be factorizable.}

\item{clusters}{a list of vectors : the variable index per clusters.}
}
\value{
A matrix on size the number of clusters (i.e. \eqn{K}). It
corresponds to the reduced matrix \eqn{R} of the original matrix.
}
\description{
The block matrix models is defined from two elements : the cluster's partition
\eqn{\{C_1, \dots, C_K\}}, included in \eqn{V} and the \eqn{R} matrix which belongs
to \eqn{\mathcal S_K(\mathbb R)}, the set of symmetric \eqn{K \times K} matrix.
The expression of the precision \eqn{\Theta} which statisfies the block matrix model is
\deqn{
   \Theta = U R U^t + A
}
where \eqn{U} is the cluster matrix and \eqn{A} a diagonal matrix such that the rows of
theta sum to zero :
\eqn{
a_{ii} = - \sum_l p_l r_{kl}
}
for \eqn{i} in cluster \eqn{C_k}.
}
\examples{
# We can build a Theta which is a block matrix as described upper :
Theta <- matrix(
 c(4.5, .5, .5, .5, -2, -2, -2,
   .5, 4.5, .5, .5, -2, -2, -2,
   .5, .5, 4.5, .5, -2, -2, -2,
   .5, .5, .5, 4.5, -2, -2, -2,
   -2, -2, -2, -2, 6, 1, 1,
   -2, -2, -2, -2, 1, 6, 1,
   -2, -2, -2, -2, 1, 1, 6),
 nc = 7
)

clusters <- list(c(1,2,3,4), c(5,6,7))

extract_R_matrix(Theta, clusters)

}
