% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR_clusterpath.R
\name{hr-clusterpath}
\alias{hr-clusterpath}
\alias{HR_Clusterpath}
\title{Clusterpath algorithm for Hüsler-Reiss models}
\usage{
HR_Clusterpath(
  data,
  lambda,
  zeta,
  W = NULL,
  p = NULL,
  eps_f = NULL,
  kappa = 0.01,
  eps_conv = 1e-07,
  tol_opt = 0.001,
  iter_max = 1000
)
}
\arguments{
\item{data}{A \eqn{n \times d} matrix : the matrix of data.}

\item{lambda}{If a positive number, the regularized parameter for the penalty. If a numerix
vector of positive number : the grid line for \eqn{\lambda}.}

\item{zeta}{A positive number : tuned parameter for the exponential weights.
If \code{W} is not \code{NULL}, this parameter is not used.}

\item{W}{A \eqn{d \times d} matrix of positive number : the weights for the penalty.
If \code{NULL} (default), it is set as the exponential weights, which depends of only one
tuning parameter \eqn{\zeta} and where we have some theoretical results.}

\item{p}{A numeric between 0 and 1, or \code{NULL}.  If \code{NULL} (default), it is
assumed that the data are already on multivariate Pareto scale. Else, \code{p} is used as the probability
in the function \code{data2mpareto()} to standardize the data (see \code{graphicalExtremes} documentation).}

\item{eps_f}{A positive number : tolerance threshold for merging clusters. If \code{NULL} (default),
it is set as \eqn{\kappa} times the median of the non-diagonal elements of the distance matrix
between the column of the initial \eqn{\Theta}.}

\item{kappa}{A positive number : tuned parameter for the fusion threshold. If \code{eps_f} is not \code{NULL},
this parameter is not used.}

\item{eps_conv}{A positive number : tolerance threshold for the convergence of the algorithm.}

\item{tol_opt}{A positive number : tolerance for the optimal step in the gradient descent.}

\item{iter_max}{An integer : maximal number of iterations of the procedure.}
}
\value{
\code{HR_Clusterpath()} is a block gradient descent from the data with default values and method for
the optimization : the variogram matrix \eqn{\Gamma} is the empirical variogram and the weights are
set as the exponential weights, which depends of only one tuning parameter \eqn{\zeta} and where
some theoretical guarantees are provided.

The output is an object of class \code{HR_Clusterpath} which contains a list of several datas :
\itemize{
\item $results : a list of results for each value of lambda, where each result is a
list with the following elements : $R the \eqn{R} matrix of the clusters, $clusters
a list of the variable indices separated per cluster, $likelihood the value of the
penalised negative loglikelihood and $lambda the corresponding regularized parameter.
\item $Gamma : the variogram matrix used for the procedure.
\item $inputs : a list of the input parameters used in the procedure.
}
}
\description{
Block gradient descent based on the Clusterpath algorithm for Hüsler-Reiss
graphical model. Usefull when the precision matrix \eqn{\Theta}
(or the variogram \eqn{\Gamma}) has a block matrix structure.
}
\details{
The block matrix models are characterized by two objects : a clustering partition
\eqn{\{C_1, \dots, C_K\}} of \eqn{V = \{1, \dots, d\}} and a matrix \eqn{R}
which belongs to \eqn{\mathcal S_K(\mathbb R)}, the set of symmetric \eqn{K \times K}
matrix (See also \code{\link{build_theta}()} and \code{\link{extract_R_matrix}()}).

The Clusterpath aims to find the optimum of some penalised negative loglikelihood defined in
\code{\link{neg_likelihood_pen}()}.
}
\section{Some results for \code{HR_Clusterpath()}}{


When we get replications \eqn{X_1, \dots, X_n} of a random vector \eqn{X} which
belongs to the attraction domain of a \eqn{K}-block Hüsler-Reiss graphical model, the
minimum of the penalised negative loglikelihood conveges almost surely to the
true precision matrix of the model \eqn{\Theta^*}, for all \eqn{\lambda}, provided that
we choose a sequence \eqn{(\zeta_n)_{n\in \mathbb N^*}} which grows slower than the
\eqn{log(n)^s} sequence with \eqn{s < 8}.
}

\examples{
# Initialization of the parameters for the simulation
R <- matrix(c(1, -3, 0,
              -3, 2, -2,
              0, -2, 1), nc = 3)
clusters <- list(1:5, 6:10, 11:15)

# Building of the variogram matrix for the simulation
Theta <- build_Theta(R, clusters)
Gamma <- graphicalExtremes::Theta2Gamma(Theta)

# Simulation of data
set.seed(123)
n <- 1e4                               # Sample size
d <- ncol(Gamma)                       # Number of variables
data <- graphicalExtremes::rmstable(
           n = n,
           model = "HR",
           d = d,
           par = Gamma)

# Computation of the multivariate Pareto data
norm_inf <- apply(abs(data), 1, max)
quantile <- sort(norm_inf, decreasing = TRUE) |> as.numeric()
k <- floor(0.1 * nrow(data))
u <- quantile[k]

data_par <- data[norm_inf > u, ] / u         # Multivariate Pareto data

# Computation of the Hüsler-Reiss Clusterpath
zeta <- log(n) ** 2              # Zeta parameter for the exponential weights
HRC <- HR_Clusterpath(data_par, lambda = 10, zeta = zeta)
HRC
}
