% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{HR_Clusterpath}
\alias{HR_Clusterpath}
\title{Optimization results from data}
\usage{
HR_Clusterpath(data, zeta, lambda, eps_g = 0.001, eps_f = 0.01, it_max = 1000)
}
\arguments{
\item{data}{\eqn{n /times d} matrix : the data.}

\item{zeta}{a positive number : tuned parameter for the exponential weights.}

\item{eps_g}{d}

\item{eps_f}{d}

\item{it_max}{d}

\item{lamda}{a numerix vector of positive number : the grid line for \eqn{\lambda}.}
}
\value{
Returns the optimization path results from \code{get_clusters()} for each
\eqn{\lambda} from the data.
}
\description{
Optimization results from data
}
